Fix GCC path for UI tests targeting Bazel
Currently, when a Bazel project is prepared for import during
a UI test, the CC environment variable is explicitly set to a path
which does not point to a gcc binary (whoops!). This causes sync
failures in UI tests which work with Bazel projects that have native
code.

This CL resolves the issue by updating the gcc path to point to
a prebuilt gcc binary. This is currently hardcoded to a Linux x86
binary because a Mac-compatible binary isn't available in prebuilts
(this shouldn't make a difference, since all our target Bazel UI
tests are run with the no_mac tag anyway). If we plan on running
these UI tests on Mac machines in the future, we will need to add a
prebuilt Mac gcc binary and update the getGccPath() logic.

Bug: 73358558
Test: Manually confirmed that SimpleBazelApplication syncs successfully
      after cherry-picking Iee60b860
Change-Id: I17fbc921005507182f966fcf8eb947ff208a85f1