Reordering navigation NlModel listeners
Previously DestinationList was listening to modelChanged, which is
itself is (currently) responsible for actually updating the NlComponents
to match the underlying Psi. The DestinationList listener might get run
before that update happens, and so would see outdated information.
Updated so that modelDerivedDataChanged is called after the updated is
done (as it is in the layout editor), and DestinationList listens on
that.

In the future we should fix it so that the updating of the NlModel is
done by the model itself, and happens before the listeners fire.

Change-Id: I0333ecad6d0147b8f4542b2193c660c483c50078
Fixes: 79441497
Test: DestinationListTest