Restart IDE if WelcomeFrame not showing in UI test
Previously, at the start of a test we would assert that only the welcome
frame was showing, and skip the test if not. This led to many tests
being skipped if the IDE got into a bad state.

This change replaces this assertion with an IDE restart so that
subsequent tests won't be skipped.

Bug: 79780751
Test: verified manually
Test-Strategy: ui.test.groups=TEST_FRAMEWORK
Change-Id: I44cd81a613048dd782e5ab8526b1f20f0476a43c