Don't warn when no content roots are found
This CL deletes a LOG.warn statement that warns when no content entries
are found in a module. Due to various complexities in mapping a Blaze
project into IntelliJ, such a scenario is somewhat common currently for
a Blaze project, so we end up triggering this warning all the time -
enough for it to swamp idea.log files and show up in some freeze traces.

Test: N/A
Bug: none
Change-Id: I8daf896e7c8413760acd2399ed0d277fe034bfa7