Check if Gradle is used before suggesting to use it on legacy projects
When a legacy Android Studio project is imported we create a
notification that suggest the users to use Gradle, even if the project
is in fact using it. Also, when importing a legacy projec, sync is not
performed and that causes checkForSupportedModules to display false
errors.

This CL checks if the project is built using Gradle and show the
notification accordingly. Also, move checks of unsupported modules only
if sync was not requested.

Test: Confirmed manually that no notification was shown for legacy
Gradle projects and confirmed that checkForSupportedModules is called
whn sync is successful.

Bug: 77864907
Change-Id: Ib718fa28bb9ab5351e2b2a779b5baf1dd6c2cb85