Fix caching of (non)namespaced repositories
- Make ResourceFolderRegistry into a project service that keeps a weak
cache of ResourceFolderRepository instances, both namespaced and not.
This way we only keep in memory what's actually being used by
ResourceRepositoryManagers.

- Guarantee that the repo manager's state corresponds to the namespacing
mode of the module by using different instances for different
namespacing modes. In practice only one instance is stored in the facet,
the other (if it exists in the first place) will be garbage collected.
There is no data that can be shared between the modes and considering
some non-obvious locking used by ResourceRepositoryManager, it seems
easiest to just let go of the whole old manager instance rather than
trying to atomically update it.

Bug: 78765120
Test: ResourceFolderRepositoryTest, MigrateToResourceNamespacesProcessorTest
Change-Id: Ic622a3e82209a72b66c920f93bb8275a114fb9a0