Systrace can return threads with the same name, resulting in invalid main.
Some captures have multiple threads with the same name, this results in
the wrong main thread being selected when the capture is selected.
As a fix to this, the CpuThreadInfo for the capture trees now set
the process id and process name. The CpuCapture constructor now
checks for this and will use the entry where the thread id matches
the process id as the main thread. If one is found.

Test: Updated To check for duplicated thread names
Bug: 77863343
Change-Id: Ibdd8b372e130acaf6e4c22d1b6016257a7006133