Removes updater dependecies so as to avoid redrawing unnecessarily
This CL removes dependencies on Updater when the dependency should be on the
dependent Range.

It also fixes various redraw issues that were hidden due to the always-redraw
behavior.

Energy axis formatter also had a constructive interference removed (where the
axis could lerp to infinity). Fix for the old behavior was to round up the
return value for getMajorInterval on the fail case, so as to ensure the major
interval is at least as large as the range (issue arose from double precision
rounding). Also, now the max label is capped at the largest label's range. This
*may* also have fixed some outstanding issues with Energy profiler's axis.

Change-Id: I15178b7e2e848327a15e01f3fe50431c603380b7
Fixes: 79556217
Fixes: 79990819
Fixes: 79431662
Bug: 79475615
Bug: 75032892
Test: Updated/added tests for new behavior