Use project-wide listener to invalidate sample data
Currently, SampleDataResourceRepository creates one VirtualFileListener
and one PsiTreeChangeListener per module for which a resource file has
been opened this session. These listeners are responsible for
invalidating SampleDataResourceRepository instances when sample data
files are modified.

In the interest of keeping the number of BulkFileListener instances
to a minimum, this CL cuts the number of listeners down to a single
instance of the new SampleDataListener class per project. Callers can
ensure this listener is registered to listen for VFS changes through the
ensureSubscribed() function. The listener is notified about PSI changes
by the already-existing PsiProjectListener component.

Bug: 79390264
Test: SampleDataResourceRepositoryTest#testSampleDataFileInvalidation
      + ResourceFolderRepositoryTest

Change-Id: I44a107bcaf81b10882ba95cdf48b26eca2faf98a