Reading marker timestamp from atrace file.
This change updates how we read the parent_ts attribute from atrace
captures. Before we assumed that the start of the capture was a
parent_ts marker that helped us align our data. However this was
off by a few nanoseconds. Now captures recorded by studio have a
parent_ts marker written just before we stop the trace. As such
we take the delta between the BOOTTIME and MONOTONIC clocks then use
that as our alignment between atrace captures and our timeline.

If a parent_ts marker is not found we use the beginning of the trace
as our start time. This allows us to start imported captures at 0s.

Test: Parser test have bene updated, trebuchet test TBD in trebuchet
branch.
Bug: 78191316

Change-Id: I13d69d133702d1fb5f2b37bc7e2efe74764488d5