Fixing idle time rendering. Now shows properly when end is off screen.
Before the idle time was using the clamped width of the drawing area
to define how to draw. Now it uses the full drawing area to define
its bounds and the clipped drawing area to clip itself to the window.

Note: The clipping was left in because this can provide significant
performance improvements.

Test: Updated CaptureHRendererTest to validate clipping bounds are used.
Change-Id: I637381d9ca25dbe8a238c7e4199d2e807eda79e7
Fixes: 79442146