Reset the presentation cache after action update.
The ActionsToolbar has a toolbar with a dynamic set of actions.
The toolbar has a Presentation cached per action. After we
replace the actions on the dynamic toolbar the existing elements
in the presentation cache are useless. And in fact this cache
keeps growing as the user navigates the components on the form.

This CL clears the presentation cache after the actions have been
updated to avoid keeping unused elements in the cache.

Bug: 79110899
Test: Added a unit test.
Change-Id: I39e957326209bf1a4e2ca79ed72561ebc07079cf