Fix the issue about not starting startup profiling
Startup profiling was not starting in some occasions,
because Studio would start and end a session for
the alive pid (or pid restored by emulator snapshot) that is going to be
killed, in a result Studio Profiler clears all startup profiling related
data from profiler. So, the new pid will be launched with the ongoing
startup profiling in the background, however our profilers would not
know about it.

Test: 1. Start startup profiling on an emulator
      2. Stop the ongoing profiling
      3. Close the emulator
      4. Start startup profiling again on the same emulator
      5. With this CL startup profiling starts as expected (previously
      it was not).
Fixes: 80174844
Bug: 76426618 # Part of
Bug: 79938546
Bug: 79484148
Change-Id: Idf521b191b819d3aef91e0de707eea0f6d86dc44