Add duration to Kernel tooltip in systrace captures
Add thread slice duration to CPU kernel tooltips in systrace captures.

As a collateral effect, CpuCapture and CpuThreadInfo were refactored so
CpuThreadInfo is responsible to determining if a thread is the main thread
or not. This way, CpuCapture doesn't need to rely on the thread name to
determine if it's the main one.

Bug: 80175064
Test: Added duration to CpuKernelTooltipViewTest#textUpdateOnRangeChange.
      Also updated tests affected by refactoring.
Change-Id: I884da71784d19f957b83df4741545605ef2e379c