Remove AndroidModuleSystem.getDependencies
Currently AndroidModuleSystem.getDependencies is only used in
DependencyManagement.kt to detect whether or not the module depends on
androidx || old support lib.  This CL changes the implementation to use
existing APIs offered by AndroidModuleSystem that no longer needs to
walk through all of the module's dependencies, which can become very
expensive.

getDependencies is removed because it has no real use-cases where it
can't be implemented with existing project systems APIs in a more
effective manner.

This CL is a part of the migration effort from old dependency management
APIs to the newer, more flexible ones.  This CL only does in-place swaps
and does not make logical changes unless stated above.

Bug: 77874749
Test: No dependency management tests are affected. All related tests
pass locally.

Change-Id: I657725d8eba1095823a0bcd80637c8b096b09777