Refactoring: Remove addChild from InspectorLineModel.
Remove addChild from InspectorLineModel and allow the
InspectorBuilders to specify a parent when a line is added to the
InspectorPanel.

The previous API required a child line to be added to the panel and
the parent in 2 separate calls. This API allows the same to happen
in 1 call.

This CL is a preparation for handling the tricky separators on each
side of a title. There are no functional changes in this CL. This is
a pure refactoring.

Bug: 73778668
Test: Existing tests should pass.
Change-Id: I9f852dbe7c9d3e94a93729770173076f413c4c60