Redo the inspector separator logic.
A title line according to a design by csinco@ has a dark background
with border lines top & bottom. A blank separator is added between
the title and elements in the inspector.

This was first implemented in ag/4034019. However this implementation
did not handle the following cases:
- tables should not be preceded by blank separators.
- need a blank separator if the first element is not a title or table.
- in addition nested expandable editors were not handled correctly.

See the added test for how this is supposed to work.

Bug: 73778668
Test: Added test for the correctness of this solution.
Change-Id: I61503983de318e1632c83482bf12c7b3365f2e1d