[AVD Mgr] Command-line AVD Mgr creates AVDs like the UI version
An AVD created by the stand-alone command-line AVD Manager had
different settings than an AVD created by the GUI AVD Manager.

This CL makes the two AVDs more similar. (But not identical.
The use cases and controls are different.)

Some classes and methods had to be moved to allow the two versions
to share code.

Fixes: 79926458
Test: Updated StorageTest, AvdManagerCliTest. AvdManagerTest still passes.
Change-Id: Ic93488eab18f750fb6df2929900b88c536e987a2