[Gradle Build Model] Fix adding dependencies with variables
This change fixes setting dependencies by ArtifactDependencySpec
by ensuring that the whole Gradle coordinate is a GString when
any one of the parts is a variable.

This change also changes anchoring to try and find the last leaf
of the anchor tree when its PsiElement is null. This can happen
when blocks are not used i.e ext.prop = ...

Test: Unit test added
Bug: 67385013
Change-Id: Id5f30ce495eff1b2532815529f4199ab3b3701cc