Adjust the left/right margins of CommonBorder.
The size of the area painted was inverted because the padding was
specified as 0. The padding is now adjusted with a minimum value
from the corner radius.
In addition the right to left drawing has been fixed.

CommonComboBox and CommonTextField were registered as opaque.
This caused the corners to be painted with the background color
which is undesired. This has been fixed.

This CL also removes the explicit background of MarginWidget.

And change the font used, and limit the combobox tests to 1
character to minimize the font differences in the images on
different platforms.

Bug: 78458943
Test: Visual test updated.
Change-Id: If9d165eed1d507e3caaa4bbf41197a4d06bbc65f