Slice Provider Wizard: Check for Compatible compileSdkVersion
Slice Provider is available on Android 'P' and over (API >= 28)
but forcing the Template minApi/minBuildApi to be 28 does
not allow this template to be created on pre-release P.
Pre-Releases of 'P' have a minApi of 27 (will be 28 only
after 'P' is released).
This CL updates the UI to account for pre-Released APIs.

Bug: 79996779
Test: Manually inspect that "Slice Provider" can be created on 'P'
Change-Id: I904f2d7e5c14a69b699539bf22f3b23b40c78bf6