Don't throw exception in getResolvedDependency.
Currently GradleModuleSystem#getResolvedDependency throws
a DependencyManagementException when the AndroidModuleModel isn't
available. This should be one of the expected possible states and should
not be exceptional.

This CL changes getResolvedDependency to return null in order to
indicate that the dependency isn't available. This is similar to how
GradleDependencyManager behaves when the model isn't available (it
pretends the dependency isn't available).

Bug: 80530667
Test: Dependency management related tests pass. Added new test to
verify getResolvedDependency returns null when there is no
AndroidModuleModel.
Change-Id: I822ad67669aad14bd8727fa98a802a1f2f8bb066