Use fake ADB server for X86AbiSplitApksTest
This test doesn't actually need a device nor emulator for any purpose.
The test only needs the Android systems to query the architecture of the
system. The architecture is needed so the IDE knows what type of APK to
build for deployment. For example, the IDE should build an APK with x86
libraries for x86 Android systems.

Test: X86AbiSplitApksTest - ran locally
Bug: 77805073
Change-Id: I0fe48fd6ed36beec5dd7b80fcdacf537ca13aee9