Use fake ADB for AbiSplitApksTest
Similar to the X86AbiSplitApksTest, this test doesn't actually need to
use the emulator nor an Android device for any purpose other than to
tell the IDE what architecture the Android system runs on. This
functionality is very easily replaced with the fake ADB server and a
custom command handler for the "getprop" command.

Bug: 77805073
Test: AbiSplitApksTest, X86AbiSplitApksTest - ran locally
Change-Id: If45fa307cde8276474390ab99fb16703e583a068