LayoutPsiPullParser errors when project disposed
In unit tests we sometimes run layoutlib asynchronously, and it would
fail if the test had already completed. This would lead to the crash
reporter being kicked off, and then the crash reporter thread would show
as leaked.
Checking for whether the project is already disposed before trying to
get the root psi element, which is where problems were occuring.
(Note that there could still be a problem if the project is disposed in
the middle of rendering, but this is unlikely).

Test: this is it
Bug: 109861864
Change-Id: Ib9ee2736d8e1b19edb590b26fa5df54ca487be02