Changes to avoid modifing BaseRefactoringProcessor
We leave the migration open for a bit longer so the helpers can still
see the information of our migration when being applied.
We also run gradle sync after we are done with the migration to avoid
the refactoring helpers removing packages because the consider them
invalid.

This change also will allow us to revert changes into the
BaseRefactoringProcessor

Test: In a separate CL
Bug: 80407667
Change-Id: I3a74d4f84b9d446584503f0e775d3b0dc9a5e6ab