Move GradleModelConverter.kt into sdk-common
This file needs to be visible to both lint and the IDE, so sdk-common
is a good place for it. It may move out of sdk-common in the future
if and only if we break the dependency of sdk-common on builder-model,
and at that point it will probably go to the same place as the rest
of the Ide* model classes.

Bug: 71554108
Test: GradleModelConverterTest
Change-Id: I4e5b69a56ccc541d8826d90caaed1c52d3a85d6f