Revert "LayoutPsiPullParser errors when project disposed"
This reverts commit ae386fe8873e62d34d35d8006dca68fff373cddc.

Reason for revert: it looks like there is a race condition which
affects random tests across the board, so the issue is indeed
unlikely to have anything to do with these tests themselves. However,
the PSQ runs history shows that re-enabling NavSceneTest led to
much more severe manifestations of these issues, so in an attempt
to unblock the PSQ, it seems reasonable to revert now until a further
action.

Change-Id: Ib9d3c27e0570997a349ba225f7a6c70311230ec9