Add a new getResValues() method to AndroidModel
Move the gradle-specific logic in
DynamicResourceValueRepository.getFullTable to a new getResValues()
method on AndroidModel, as a first step to providing this functionality
in a build-system-independent manner.

This restores the code reverted by commit
78a6d8f7e6e9d9c464c9e7512eb3aa574c9c2d37, originaly implemented in
commit 9f264c79950544224c5e7765e1326b227b479e95. However this patch
additionally includes a fix for a bug introduced in DynamicResourceValueRepository.

The original patch was applying the resource overloads in an incorrect
order, causing the flavor values to override the build type values.
That caused a test failure NlPreviewTest.testRenderingDynamicDependencies.
This patch reverses the order and fixes the failing test.

Bug: 109665898
Test: NlPreviewTest
Change-Id: Ibe73f2bc38babe564ad939e2ced75a4fddbb58d3