Increase timeout of gradle output waiting
Gradle 4.7 changes its standard output format for per-task output
(https://docs.gradle.org/4.7/release-notes.html) which makes first batch
of outputs of a task slower as it's firts written in a per-task output
stream and after passed to Gradle standard output.

This can be seen by running the test and noticing that first few lines
of the task output is written at the same time after a small interval.

Test: GradleTasksTest.testNotFinishedImmediatelyAndOutputIsShown
Bug: N/A
Change-Id: I488f2ca04d9a04f8d723c547dd9a624150db44b1