Refresh project system each ClearResourceCacheAfterFirstBuildTest case
ClearResourceCacheAfterFirstBuildTest assumes that the TestProjectSystem
starts in a fresh state at the beginning of each test case. This assumption
is false, as AndroidProjectRule uses a shared project instance for each
test case, which in turn uses a cached project system instance. This
results in test flakes that depend on the order in which the test cases
are executed.

This CL resolves the issue by mocking ProjectSystemService and having
the mock return a fresh project system for each test case.

Change-Id: I1121a6ef60058629ad34b64c8432f2508b2829dc
Fixes: 80590520
Test: ClearResourceCacheAfterFirstBuildTest