Migrate AddRemoveCppDependencyTest into Bazel
This migration includes a workaround for b/80429739. Cmake's ninja tool
has a bug where the tool can not process files that are more than 32
directories deep from the root of the file tree. This is a problem
that's easily encountered when running a test within the Bazel sandbox,
since Bazel will recreate the the path to a package's location from the
root of the Bazel workspace.

This workaround will persist until a newer version of Cmake is used as
the default in Gradle, preferably Cmake 3.10.

Bug: 80429739, 74079626
Test: bazel test //tools/adt/idea/android-uitests:AddRemoveCppDependencyTest
Change-Id: Iaa0932769de11216af8ba7df220d912ff0261c98