Allow PTableImpl.startNextEditor() to edit name cells.
This method only allowed keyboard navigation to value cells.
The CL will now allow editing of cells in the name column as well
if the model allows editing in these cells.

This change is a preparation for allowing adding property names
in the "Declared Attribute" section of the new property panel.

This change also adds tableModel as a property of PTable.
Also as a preparation for adding properties.

Bug: 73778668
Test: Added unit test.
Change-Id: Ia406e81894df90618eff0cfc21b03afb7bcb827f