Don't traverse whole res dir when confirming new image/vector assets
Currently, ConfirmGenerateIconsStep and ConfirmGenerateImagesStep both
use FileTreeModel to present the user with a view of the changes to the
file system that will happen as a result of generating the new image or
vector asset. Though these wizard steps only ever show a predetermined
set of relevant files, the FileTreeModel recurses through the entire
resource directory.

This CL resolves the issue by replacing FileTreeModel with
ProposedFileTreeModel, which only operates on the portion of the file
system which is relevant to a predetermined set of files. This CL also
removes FileTreeModel.java and its associated classes, since these were
its only usages.

Bug: 80198830
Test: ProposedFileTreeModelTest
Change-Id: I218d8f50024b61f89ffca71b533b85f576565a16