Show quick fix to disable offline mode on missing dependencies
Gradle will not download missing dependencies when working in offline mode.
If there is a missing dependency and off line mode is enabled, we show an
error message and an option to re try, but there is no clear indication
that offline mode is enabled.

This CL adds a quick fix to disable offline mode when missing dependencies
are reported.

Bug: 65188424
Test: New tests included
Change-Id: If3f057afdb9d8e2762bb4d7c5d3f25524c15ffb0