Use kotlin-m2repository maven repo in tests
The test_deps rule will no longer contain the kotlin gradle plugin and
its dependencies; these are moving to a separate maven repo in another
change in this topic. This change updates BUILD files to include this
new repository, and includes it in the setup done in the Suite classes
for which it's necessary.

Bug: 110231494
Test-Strategy: ui.test.groups=DEFAULT
Change-Id: Icaa61fe620b75733eeb67f4eeb8e149d372fbbb3