Adding a sync point for DataStoreTable error handler.
Because there is a seperation between the Datastore Service and
the DataStoreTables We have a callback that routes errors from
the datastore tables to the service to properly handle them.

This set gets added to each time we open a new project and remove from
each time we close a project. This means that two threads modify this list.

Test: None
Change-Id: Id95c59dfea808a1a7697a0958182ca6c0d580e14
Fixes: 80111520