Show selected row in property tables.
Update all property cell renderer to display white on blue for
the selected row in a table (when the table has focus).

To do this I had to implement a CellPanel and a CellBorder which
are capable of propagating the foreground and background color to
its sub components.

Also in this CL:
Make sure the transition to and from cell editing can be handled
entirely through the keyboard:
- enter will start cell editing
- esc will stop cell editing (excpet for an open combo box)
- space should toggle the value of a boolean editor
- general typing should start cell editing

Bug: 73778668
Test: Updated unit tests.
Change-Id: I41d6ecb18053a9445af8c2a593180acda07ffcc9