Fix race conditions in deviceChanged, clientChanged, and addTargetDevice
The methods deviceChanged and clientChanged run in separate threads. One
thread can modify the myClients list while the other thread is dependent
on the list not being modified. This can occur when the Client's
description is not updated when deviceChanged runs. Thus, deviceChanged
can believe that no Client exists. Meanwhile, the Client's name gets
updated right after deviceChanged checks for the Client name. This can
cause clientChanged to believe that a Client exists, and clientChanged
will add the Client to the myClients set. When deviceChanged later
checks the myClients set and finds that it's non-empty, deviceChanged
will believe that the application crashed on the device. deviceChanged
will proceed to prematurely end the instant run session.

A separate race condition can occur when deviceChanged accesses
myDeviceAdded variable before they are initialized in addTargetDevice.
If the myDeviceAdded variable is not yet initialized by the time
deviceChanged reads the variable, then the method may read 0 for
myDeviceAdded. myDeviceAdded is compared to the returned value from
System.currentTimeMillis(). If the read of myDeviceAdded is 0,
deviceChanged will consider the client to have "timed out", and
deviceChanged will prematurely end the instant run session.

A third threading issue is where clientChanged can race with
addTargetDevice. In this case, clientChanged may read myDevices and find
that addTargetDevice has not yet added a device to the myDevices set.
Additionally, there is no guarantee that addTargetDevice will see
up-to-date Client values for writes that occurred in the clientChanged
thread.  The result is that both methods may end up not adding a client
to the myClients list, and the client never gets added during the
instant run session.

This commit attempts to solve all three threading problems by using a
lock that synchronizes access to the myDevices, myClients, and
myDeviceAdded variables. This ensures that access to the myDevices,
myClients, and myDeviceAdded states are atomic, and that writes in one
thread will be visible to another thread.

Change-Id: I3dd57fe7c25aaaecdb71f18f84ce0b0a33bbe155
Fixes: 110490227
Test: bazel test //tools/adt/idea/android-uitests:ChangeManifestTest
--runs_per_test=100