Always use classpath JAR in UI tests
Recently one builder running UI tests with ant started hitting the
command line length limit. The others (with a slightly shorter path
prefix) must not be far behind.

This problem was already addressed for running UI tests in Bazel, by
building a JAR that contains only a manifest specifying the classpath,
and using that as the argument to -classpath on the command line.

This change makes it so this classpath jar is built and used regardless
of build system. For running on the release, there is no change, as
there is no classpath to specify (at least at this level).

Also clean up some classpath construction logic that was inherited from
JetBrains and seems to not be needed.

Bug: 110530437
Change-Id: Idee2bd80f2f31e184cfff0e8c91049fd5763256b
Test-Strategy: ui.test.groups=THEME,EDITING,PROJECT_SUPPORT,PROJECT_WIZARD,TEST_FRAMEWORK,DEFAULT