Limit the scenario where the profilers keeps the same process
Previously, the profiler would select and profile the same process
as long as the pid, name and start time were identical. This change
further limits that by only selecting the process iff its state has
not changed, or it went from alive to dead.

This helps to prevent the case where the profiler would start a new
session on the same process when a device is connected or during
emulator snapshot boot up.

Bug: 77658332
Test: StudioProfilersTest#testSessionDoesNotAutoStartOnSameProcess
Change-Id: I3cfa7e038aa936f29ffbad400e2a90be130fc5d9