Syncs project when new project from VCS
When creating a new project from VCS in File -> new project, AS was just
opening the project without importing it, which causes missing
fonctionalities.

This CL fixes this issue by checking if a Gradle project already has
Gradle facets or not when opening, and in case it doesn't, run sync.

Bug: 67668533
Test: unit test GradleProjectInfoTest and confirmed manually for the
entire integration file -> new project from vcs
Change-Id: I4829e4ed970f9dd07b54bad00104c66a0c17ce7f