Move image scaling logic in ConfirmGenerateIconsStep to separate class
ConfirmGenerateIconsStep downscales icon previews so that they fit well
within the preview file tree. This is an easily extractable subproblem
that should be moved to a separate utility class to make it easier to
test and to de-clutter the logic in ConfirmGenerateIconsStep.

This CL resolves the issue by extracting the image scaling logic to a
new ProportionalImageScaler class.

Bug: N/A
Test: ProportionalImageScalerTest
Change-Id: Id06983a9ced7af663f5e9cfa7e5d4cdbba941ad2