Actionbar resizing incorrectly on zoom
When the zoom factor changes the width of the zoom level label changes.
Previously when we re-laid out the action bar we were only resizing the
content and not the bar itself. If the zoom label got wider, the content
would no longer fit within the previous width of the action bar and it
would wrap onto two rows. This (or maybe something else) would then
cause the parent of the action bar to be revalidated, and then it would
get resized correctly and the user wouldn't notice the brief
double-height action bar.
However, this meant that a resize event could be sent to the design
surface, which in this case caused us to zoom to fit (because of a
separate problem that will be addressed in a subsequent CL).

Fixes: 110435862
Test: DestinationListTest.testSelectComponent