Fixing current zoom
If you set the zoom to e.g. zoomToFit and then change the scale some
other way (programattically) the current zoom type shouldn't still be
set to fit.
Additionally, it doesn't make much sense for the current zoom type to be
"in" or "out," so making it not set those.

Test: NavDesignSurfaceTest.testSkipContentResize
Bug: related to 4420651
Change-Id: I010d42023b775fb776d8fa05d9fb84afc7669817