Removing updatable from the ThreadsModel, adding listener to Kernel.
The kernel model did not have a range aspect listener. This means that
whenever the range updated without a tooltip being present the kernel
table would not redraw.

Removing updatable from ThreadsModel, and adding the aspect listener
also resolves the AxisComponent lines not showing. However I suspect
this is due to draw ordering more so than anything else.

Final note, this should show up as an improvement on the performance
screen dashboard thingy.

Test: None
Bug: 109836145
Change-Id: Iba1ee61b46f7cc6e1e5577618f72011e1193b213