[Project Upgrade] Apply deduplication to unresolved dependencies
This change extends SimpleDeduplicatingSyncIssueReporter to allow
more advanced customization and reworks UnresolvedDependeciesReporter
to make use of this. This allows deduplication of resolved dependencies.

Note: This change means that some quick fixes will only work on the
first module we find. This change should be submitted along with one
that reworks these fix as a result, for now, DO NOT SUBMIT.

Test: Tests updates
Bug: N/A
Change-Id: If5eede53f5f51b036bb68c92d75a5cc5d6a4b29f