Fixing a bug in trebuchet importer when parsing looped buffers.
A feature of systrace that was not implemented in our parser was
the ability to skip all data up to where a buffer was started.
This is needed because we may have start / end marker events that
get dropped. If this happens it means we end up in a bad state when
looking at trace markers. Aka a marker looks like it never ends.

Test: In trebuchet
Bug: None
Change-Id: I29eccc3b784eb0f1d2fa5aaf7c7b53d2adc9ad5e