Provide quickfix for MIN_SDK_VERSION_IN_MANIFEST.
A SyncIssue will be reported if minSdkVersion is defined in
manifest file. This CL provides quickfix for such issue.

If minSdkVersion is already defined in build.gradle file,
the quickfix just removes it from manifest.
If minSdkVersion is not yet defined in build.gradle file,
the quickfix removes from manifest and adds it to build.gradle.
Gradle Sync will be performed after the fix.

After removing minSdkVersion from manifest, if the uses-sdk tag
becomes empty, remove uses-sdk as well.

Bug: 75004328
Test: RemoveSdkFromManifestHyperlinkTest
Change-Id: I55dada7448bcd036db1028245533565864b60655