ResolveScopeEnlarger for light R classes
Now that light classes are not considered part of every scope, we need
to managed the resolve scope explicitly by deciding which light classes
should be visible to a given module.

To make this work, we make the module R classes have "dynamic" names,
based on the current package in the manifest (instead of creating new
classes when needed). This way there's a per-module "singleton" PsiClass
(and LightVirtualFile) that can be added to the resolve scopes and found
later.

Test: LightClassesTests.kt
Fixes: 111110952
Change-Id: Ia2f9ea9259d7988def53898f287ec1cbe5b5299f