Direct ui-test bazel output to managed tmp folder.
Currently ui-tests for bazel-based projects executes bazel on the test
machine and stores it's output files in the standard bazel output
directory (/home/<user>/.cache/bazel/_bazel_<user>/). On build bots this
directory is not cleaned regularly and will fill up rapidly as the
build bot runs a lot of tests.

This CL adds a startup option to the bazelrc of these ui-tests that
direct bazel outputs to //tools/idea/androidStudio/bazel-output, which
is cleaned regularly.

Fixes: 111261271
Test: Verified that running bazel ui tests no longer create output in
the standard directory and into the desired output directory instead.

Change-Id: I5354df6678c6f561a9bfb123efb95c6e3323812d