Use estimate duration to track length of failed captures
We need to track the length of failed captures and right now we're just
setting the capture duration to -1 if the capture is not successful. We
could use this field instead to track the approximate capture length
calculated from the difference of the device time between the user clicking
on Start and Stop.

Bug: 111298929
Test: Updated affected tests in CpuProfilerStageTest
Change-Id: I3c646d0344d795d9c4baa9e7cefd676983074505