Don't track failed captures metadata fields as -1
We shouldn't set the fields record_duration_ms, parsing_time_ms,
trace_file_size_bytes and capture_duration_ms to -1 in CpuCaptureMetadata
for failed captures. When collecting usage states, -1, which is represented
as 0xFF...FF, will show up as a long positive number and might confuse our
reports.

As we already have the capture status, we should let these fields unset, so
they will default to 0 and that's a reasonable value to have in the reports.

Bug: 111219960
Test: updated affected tests in CpuProfilerStageTest
Change-Id: Id8923900052613d7cd876d6032d76bdf967107bb