New TimelineTable, which renders a timeline axis in the header
Currently, we have been jamming the timeline axis into the first row,
but this has produced some situations where the data is hard to read.

Now, we actually create a table which reserves space in its header and
renders the timeline axis there. We also move some of the logic for
rendering cells into the class as well, to allow the header and cell
markers to lineup.

Test: N/A, visual refactoring only; otherwise, covered by existing
tests.

Change-Id: Ic8067924edee2703e9843bbe86881968de290ee6
Fixes: 37664504
Fixes: 70387425